perf(dynamic-chunk): increase chunks time limits
