+++ src/uploadx/lib/utils.ts
@@ -75,31 +75,6 @@ export const b64 = {
   }
 };
 
-/**
- * Adaptive chunk size
- */
-export class DynamicChunk {
-  /** Maximum chunk size in bytes */
-  static maxSize = Number.MAX_SAFE_INTEGER;
-  /** Minimum chunk size in bytes */
-  static minSize = 1024 * 256;
-  /** Initial chunk size in bytes */
-  static size = 4096 * 256;
-  static minChunkTime = 2;
-  static maxChunkTime = 8;
-
-  static scale(throughput: number): number {
-    const elapsedTime = DynamicChunk.size / throughput;
-    if (elapsedTime < DynamicChunk.minChunkTime) {
-      DynamicChunk.size = Math.min(DynamicChunk.maxSize, DynamicChunk.size * 2);
-    }
-    if (elapsedTime > DynamicChunk.maxChunkTime) {
-      DynamicChunk.size = Math.max(DynamicChunk.minSize, DynamicChunk.size / 2);
-    }
-    return DynamicChunk.size;
-  }
-}
-
 export function isIOS(): boolean {
   return /iPad|iPhone|iPod/.test(navigator.platform)
     ? true
